# cmp: print verbose diffs as we find them